annotateMFIHeatmap

Annotate an unmerged heatmap of MFIs with each metacluster's final assignment.

Usage

annotateMFIHeatmap(
  merged_input,
  original_input = NULL,
  cols_to_use = NULL,
  ...
)

Arguments

merged_input

The final table or FlowSOM object, whose metaclusters have been merged and named via either editTableMetaclusters() or FlowSOM::UpdateMetaclusters().

original_input

If merged_input is a FlowSOM object, the initial FlowSOM object before merging should be provided here.

cols_to_use

A character vector specifying which markers you would like to include in the plot. Default is the markers used for clustering.

...

Additional parameters to pass to ComplexHeatmap::Heatmap().

Value

An annotated heatmap generated by ComplexHeatmap::Heatmap().

Details

Additional details...