clusterSubsetWithPCA

Cluster a subset of flow data using principal components.

Usage

clusterSubsetWithPCA(
  input,
  pca_obj,
  num_components,
  num_clus,
  xdim = 10,
  ydim = 10,
  seed = NULL,
  convert_to_channels = TRUE,
  fsom_file = NULL
)

Arguments

input

A FlowSOM object or data frame.

pca_obj

A prcomp object as generated by doPCA()

num_components

The number of principal components to use for reclustering, as determined by examining results of plotPCAScree().

num_clus

The number of metaclusters you would like.

xdim

x-dimension of the SOM. Default is 10.

ydim

y-dimension of the SOM. Default is 10.

seed

A seed for reproducible results. Default is NULL.

convert_to_channels

Logical, should the returned/saved FlowSOM object's data be defined in terms of channels. Default is TRUE.

fsom_file

An .rds filename. If not NULL (default), the resulting FlowSOM object will be saved under this name.

Value

A FlowSOM object or data table, depending on the data type of input.

Details

When reclustering a subset of flow data, such as a cell population that has been identified as monocytes, not every marker you used for the initial clustering is still relevant. In such a case, you would no longer be interested in using, for instance, a B cell marker. One solution is to leave it up to the researcher to determine what markers are appropriate for the reclustering, but this approach is very user-dependent, therefore lacking reproducibility.

This function solves this problem by using PCA for reclustering. It takes in an aggregated dataset, as generated by createFilteredAggregate(), and clusters the cells within it using the specified number of principal components. If convert_to_channels is TRUE (default), it then takes the cluster and metacluster assignments found from this clustering to create a FlowSOM object defined in terms of channels, rather than principal components, so that it may still be used with both the functions in flowFun and in FlowSOM.