createFilteredAggregate

Filter preprocessed data to create a new aggregate file.

Usage

createFilteredAggregate(
  input,
  num_cells,
  clusters = NULL,
  metaclusters = NULL,
  dir_save = NULL
)

Arguments

input

A FlowSOM object as generated by FlowSOM(), or a data frame.

num_cells

The total number of cells to include in the aggregate file.

clusters

A numeric vector listing the clusters of interest.

metaclusters

A character vector listing the metaclusters of interest.

dir_save

If a directory path is given, then the filtered .fcs files used to create the aggregate file will be saved there. If NULL (default), then they will be discarded.

Value

An aggregated flowFrame or data table, where only cells belonging to the given metaclusters and/or clusters are included.

Details

It is only necessary to define one of the arguments clusters and metaclusters, but the user may use both if they wish.

This function takes in either a data frame or FlowSOM object, and subsets each sample to contain only the cells included in the given metaclusters and/or clusters. If desired, new .fcs files for each of these subsetted samples are then written to a new directory, whose name is specified by the user with the dir_save parameter. Note that it is only helpful to save the files if you intend to perform additional reclustering on them, otherwise they may be safely deleted.