doDAAnalysis

Do differential abundance analysis on the data.

Usage

doDAAnalysis(
  design,
  counts,
  contrasts,
  sample_df,
  norm_method = c("TMM", "TMMwsp", "RLE", "upperquartile", "none")
)

Arguments

design

A design matrix as generated by makeDesignMatrix().

counts

A count matrix as generated by makeCountMatrix().

contrasts

A contrasts matrix as generated by makeContrastsMatrix().

sample_df

A data frame, generated either manually or by prepareSampleInfo().

norm_method

The normalization method to be used and passed to edgeR::calcNormFactors(), if any.

Details

This function performs tests for differential abundances between groups of interest defined by a contrasts matrix. Models are fit and compared via likelihood ratio tests using functions from edgeR. Results are saved as a .csv file for each comparison in the directory defined by dir_edger() at runtime. If running an edgeR analysis on multiple FlowSOM objects, or running the script multiple times for any reason, the user may wish to create sub-directories within dir_edger to stay organized. To do this, simply set the dir_tables parameter to a preferred sub-directory name.

See also