plotGroupMFIBars

Function to draw bar plot of MFIs or dMFIs by group.

Usage

plotGroupMFIBars(
  input,
  sample_df,
  comparison,
  meta_to_plot = NULL,
  upper_lim = NULL
)

Arguments

input

A matrix or data.frame where rows are samples, columns are clusters, and values are MFIs for a channel

sample_df

A factor data frame as generated by prepareSampleInfo().

comparison

A list defining a comparison.

meta_to_plot

A vector of strings defining the metaclusters to include in the plot. Default is all metaclusters.

upper_lim

The y-axis upper limit.

Value

A bar plot drawn with ggplot2.

Details

The names and order of the groups defined in comparison should appear in the same way you would like them to appear in the bar plot. Make sure that the group names you supply are the same as they appear in the sample_df object.