plotGroupUMAPs

Plot colored UMAPs for each group of interest.

Usage

plotGroupUMAPs(
  input,
  sample_df,
  comparison,
  umap = NULL,
  color_by = "density",
  num_cells = 5000,
  seed = NULL
)

Arguments

input

A data.table or FlowSOM object.

sample_df

A factor data frame as generated by prepareSampleInfo().

comparison

A list specifying the groups of interest, in terms of file names.

umap

A UMAP plot as generated by plotUMAP(). If NULL (default), a new UMAP will be generated for the plots.

color_by

A string specifying how you would like the UMAP plot to be colored. Your options are "density" or a marker or channel of interest. Default is "density".

num_cells

The number of cells you would like to be sampled for each group. Default is 5000.

seed

Optional, a seed for reproducibility.

Value

Plots faceted by group drawn with ggplot2.

Details

If any group has fewer than num_cells, then the number of cells to use for plotting will be taken from the group with the lowest cell count.